Initial Prebuilt (1.34.2)
The bootstrap/ directory contains the procedures and logs for producing
the initial 1.34.2 version of the compiler. Special care was taken to
ensure that our initial prebuilt was resistant to trusting trust attacks
by bootstrapping through mrustc. See bootstrap/README.md and the
included scripts for further details.

The linux-x86/1.34.2 directory contains the initial version of the rust
toolchain. All future versions (1.35.x+) will be built in isolation on
Android build infrastructure.

Bug: 129701460
Change-Id: Icfa4d39a2735c669690ed76e93dad23a9b240d4f