rustc-1.39.0 Build 6100717
This rust prebuilt is only intended to help roll us forwards to 1.40,
and so does not include an update to the Android.bp file to use the new
prebuilts or build a new sysroot.

As rustc-1.38.0 is was not used to build Android, we remove it at the
same time.

Bug: 146571186
Change-Id: If1ed218db1283ab4be5fb33c1896542d818a7af8