libstd: link 'signal' at compilation for Android.
libstd links 'signal' for Android specifically at runtime since API
levels prior to API 20 did not have 'signal' and used 'bsd_signal'
instead. This prevents static compilation as the function is loaded at
runtime using dlsym.

Because this toolchain is building for platform, we know that our libc
uses 'signal' and not 'bsd_signal'. This means we can safely remove this
behavior and just link 'signal' in normally. This allows us to create
static binaries.

Bug: 169434439
Test: Rust static binary works.
Change-Id: I2829d7cb36dd54698458b9845ef33931df82c5e6