Populate abort message
Android has the ability to attach a message when calling abort(). Modify
the panic_abort crate to populate this message if the argument to panic!
is a String or &str. This is a temporary solution as adding `alloc` as a
dependency of `panic_abort` is unlikely to be accepted upstream.

Bug: 163343666
Test: build aosp_cf_x86_64_phone; run keystore2 without argument; inspect logcat
Change-Id: Ied9b74672212266d027f19426adaf4d6bdfc2e56