Switch rustfmt and rust-analyzer to 1.67.1 version
This works around a build issue with the latest 1.68.0 version of Rust,
because the current build rules for `rustfmt` incorrectly are treating
it as a binary that can be copied elsewhere. The build rules should
instead be invoking `rustfmt` directly from where it is present in
prebuilts/rust/linux-x86/$VERSION/. To work around the issue, we are
pinning the 1.67.1 version of `rustfmt` temporarily.

Bug: http://b/274187738
Test: m, followed by `atest .` in the Bluetooth directory
Change-Id: I0614e6bb8e561dedab13d43dfdf7a786d809281b