Don't build sanitizers for compiler-builtins.
Fuzzing builds use -C link-dead-code, which causes compiler-builtins to
not compile with rustc 1.88. The error is:

`compiler_builtins` cannot call functions through upstream monomorphizations

We are carrying a patch to work around this, but updates to
compiler-builtins make the patch hard to maintain.

We already do the same thing for libcore: aosp/1784688

Additional context in b/193200135.

Cargo fuzz no longer uses -C link-dead-code, but we aren't sure whether
we can do the same. It's not certain that we can reproduce a fuzzer
test case from 4 years ago that didn't reproduce reliably even at the
time it was discovered.

This seems like the smallest, least disruptive way to fix the compile
error.

Bug: 435702100
Test: SOONG_DISABLE_CLIPPY=true RUST_PREBUILTS_VERSION=1.88.0.test m rust
Change-Id: Ib8ed890e9bb01b583037567c59da2cd9aad9667c